+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 ### Changed
 
 ### Fixed
+- webxdc: fix a bug where the webxdc was able to open the dev tools.
+Also make opening devtools with F12 more reliable.
 
 <a id="1_40_2"></a>